ML.NN

Description

Simple neural network implementation, for learning purposes only.

Data types

type ActivationFunction = R -> R

An activation function for a neuron.

type ActivationFunctionDerivative = R -> R

The derivative of a neuron activation function.

data Layer

A layer in a neural network is a bias vector and weight matrix.

Let n be the number of neurons in the layer and m be the number of inputs to the layer. Then layerBiases is a vector in Rⁿ and layerWeights is an nxm real matrix.

Constructors

Layer

Fields

layerBiases :: Vector R

A vector in Rⁿ representing the neuron biases.

layerWeights :: Matrix R

An nxm real matrix of the neuron weights.

data Network

A network is a list of layers.

Constructors

Network

Fields

networkLayers :: [Layer]

Layers in the neural network.

Network initialization

randLayer

Arguments

:: RandomGen g

=> Int
Number of inputs into the layer.

-> Int
Number of neurons in the layer.

-> Rand g Layer
Randomly generated layer.

Generate a randomly initialized layer in a neural network.

randNetwork

Arguments

:: RandomGen g

=> [Int]
Number of neurons in each layer.

-> Rand g Network
Randomly generated network.

Generate a random neural network given the sizes of each of the layers.

For example, [3,2,4] will generate a neural network with 3 input neurons, 1 hidden layer with 2 neurons and 4 output neurons.

Running networks

feedForward

Arguments

:: ActivationFunction
Neuron activation function.

-> Vector R
Input to the layer.

-> Layer
Layer to run.

-> Vector R
Output from the layer.

Feed the output from a previous layer to the next layer.

runNetwork

Arguments

:: Network
Network to run.

-> ActivationFunction
Neuron activation function.

-> Vector R
Network input.

-> Vector R
Network output.

Run a neural network.

Training networks

data TrainingConfig

Configuration for training a network.

Constructors

TrainingConfig

Fields

trainingEpochs :: Int

Number of epochs of training to perform.

trainingEta :: R

η - Learning rate.

trainingActivation :: ActivationFunction

trainingActivationDerivative :: ActivationFunctionDerivative

trainingCostDerivative :: CostDerivative

data Sample

A training sample.

Constructors

Sample

Fields

sampleInput :: Vector R

sampleExpectedOutput :: Vector R

type CostDerivative = Vector R -> Vector R -> Vector R

A vectorized function which returns ∂Cₓ/∂a.

The first parameter is the output activation, the second parameter is the expected output.

gradientDescent

Arguments

:: TrainingConfig

-> [Sample]
Training data (input, expected output).

-> Network
Current network.

-> Network
Updated network.

Train the neural network using gradient descent.

Cost function derivatives

mse' :: CostDerivative

The derivative of the mean squared error cost function.